various additions/fixes: see comment
- added movie files into /files folder (they are actually trailers)
- various minor fixes to informer/vault/supplier
- edited the .wsdl of the vault service to let it exploit the mime attachment for the - addMovie service (so to avoid to manually edit the byte type to the DataHandler type into the generated java classes)
- added the "views" field into the movie table in the db, so to avoid to compute them every time by joining the movie table witht he user_movie one

+++ code/supplier/src/main/java/it/univaq/disim/netflics/supplier/service/SupplierServiceImpl.java
@@ -64,10 +64,13 @@ public class SupplierServiceImpl implements SupplierService {
             SupplierMovie sm = new SupplierMovie();
             sm.setSupplierId(supplierId);
             sm.setMovieId(movieRepository.findOneByImdbId(imdbId).getId());
-            supplierMovieRepository.delete(sm);
-
+            // if there is a wrong entry into the db, clean it
+            if(sm.getMovieId() != null){
+                supplierMovieRepository.delete(sm);
+            }
             return null;
         }
+        LOGGER.info("movie "+imdbId+" has been found");
         return file;
     }
 
@@ -158,7 +161,7 @@ public class SupplierServiceImpl implements SupplierService {
         } else {
             //clean up the db
             supplierMovieRepository.delete(sm);
-            throw new BusinessException("couldn't fetch movie from vault service");
+            LOGGER.error("couldn't fetch movie from vault service");
         }
 
     }
